requirements.txt
fastapi
uvicorn
pandas
numpy
pyarrow
sentence-transformers
qdrant-client
weaviate-client
redis
psycopg2-binary
neo4j
python-dotenv
beir
